[REF] web: convert TagsList tests to HOOT
This commit converts the TagsList QUnit tests to the HOOT suite.

part of task-3705027
